+++ src/main/java/dao/BookDAO.java
@@ -1,16 +1,84 @@
 package dao;
 
-public class BookDAO {
-	private static final String URL = "jdbc:mysql://localhost:3306/jdbcsample?serverTimezone=Asia/Tokyo";
-	private static final String USER = "AppUser";
-	private static final String PW = "morijyobi";
+import java.net.URI;
+import java.net.URISyntaxException;
+import java.sql.Connection;
+import java.sql.DriverManager;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.util.ArrayList;
+import java.util.List;
 
-	private static void loadDriver() {
+import dto.BookDTO;
+
+public class BookDAO {
+	private static Connection getConnection() throws URISyntaxException, SQLException {
+		
 		try {
-			// JDBCドライバのロード
-			Class.forName("com.mysql.cj.jdbc.Driver");
+			Class.forName("org.postgresql.Driver");
 		} catch (ClassNotFoundException e) {
 			e.printStackTrace();
 		}
+		URI dbUri = new URI(System.getenv("DATABASE_URL"));
+	    String username = dbUri.getUserInfo().split(":")[0];
+	    String password = dbUri.getUserInfo().split(":")[1];
+	    String dbUrl = "jdbc:postgresql://" + dbUri.getHost() + ':' + dbUri.getPort() + dbUri.getPath();
+	   
+	
+	    return DriverManager.getConnection(dbUrl, username, password);
 	}
+	
+	// 引数の Student インスタンスを元にデータを1件INSERTするメソッド
+	public static int registerBook(BookDTO book) {
+		String sql = "INSERT INTO book(id,name,isbn) VALUES( ?, ?, ?)";
+		int result = 0;
+		
+		try (
+				Connection con = getConnection();
+				PreparedStatement pstmt = con.prepareStatement(sql);
+				){
+			pstmt.setInt(1,book.getId());
+			pstmt.setString(2, book.getName());
+			pstmt.setInt(3,book.getIsbn());
+			
+			result = pstmt.executeUpdate();
+		} catch (SQLException e) {
+			e.printStackTrace();
+		} catch (URISyntaxException e) {
+			e.printStackTrace();
+		} finally {
+			System.out.println(result + "件更新しました。");
+		}
+		return result;
+	}
+	
+public static List<BookDTO> selectAlllibrary() {
+		
+		List<BookDTO> result = new ArrayList<>();
+		String sql = "SELECT * FROM book";
+		
+		try (
+				Connection con = getConnection();
+				PreparedStatement pstmt = con.prepareStatement(sql);
+				){
+			try (ResultSet rs = pstmt.executeQuery()){
+				while(rs.next()) {
+					int id=rs.getInt("id");
+					String name=rs.getString("name");
+					int isbn=rs.getInt("isbn");
+					BookDTO employee=new BookDTO(id,name,isbn );
+					
+					
+					result.add(employee);
+				}
+			}
+		} catch (SQLException e) {
+			e.printStackTrace();
+		} catch (URISyntaxException e) {
+			e.printStackTrace();
+		}
+		return result;
+	}
+	
 }
